requirements.txt
dash
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
pandas
mysqlclient
python-dotenv
MySQLdb